Revert "Made spin return an int"
This reverts commit 89b741d0ef8714df975fd8107a60ded5c4290e1a.

+++ src/man/noggin/CombinationObjects.cpp
@@ -110,11 +110,11 @@ namespace noggin {
     }
 
     // Returns -1, 0, 1 to tell the robot which way to spin
-    const int RobotLocation::spinDirToPoint(Location& other)
+    const float RobotLocation::spinDirToPoint(Location& other)
     {
         float targetH = getRelativeBearing(other);
         if(targetH == 0) return 0;
-        else return int(NBMath::sign(targetH));
+        else return NBMath::sign(targetH);
     }
